// src/store/main/system/type.ts
export interface ISystemType {
  totalQueryInfo: any
  usersList: any[]
  usersCount: number
  roleList: any[]
  roleCount: number
  departmentList: any[]
  departmentCount: number
  menuList: any[]
  menuCount: number
  goodsList: any[]
  goodsCount: number
  categoryList: any[]
  categoryCount: number
}